requirements.txt
discord.py==2.3.2
azure-mgmt-compute==30.6.0
azure-identity==1.16.0